Removing trixie for now
It appears that multistrap might have been removed
in trixie, so might have to change the build steps for
trixie, so not doing it for now - as we're not actively using
trixie based build images anywhere for now.

This should be addressed later when we do need to move to trixie.

Signed-off-by: Asutosh <asutosh@tyk.io>